backend/requirements.txt
fastapi
uvicorn
openai
chromadb
requests
pydantic
python-dotenv
